GP-45047 Automatically retry lookup with postal code
This changes the lookup approach to automatically retry using the
postal code when ambiguous results are returned.

When multiple contacts share the same name and date of birth,
postal code can be used to get a unique match. However, we
don't want to include the postal code with every lookup since
we may have an incorrect one that would prevent us from getting
any matches. With this implemnetation, we only provide the postal
code if needed.